Split out new.rs into permanent error modules.
