+++ deepwell/src/error/legacy.rs
@@ -1,5 +1,5 @@
 /*
- * error/old.rs
+ * error/legacy.rs
  *
  * DEEPWELL - Wikijump API provider and database manager
  * Copyright (C) 2019-2026 Wikijump Team
